@@ -4,7 +4,7 @@ version: 0.0.1
 author:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.10.0 <2.0.0"
 
 dependencies:
@@ -13,14 +13,14 @@ dependencies:
   flutter_web_plugins:
     sdk: flutter
   http: ^0.13.0
-  flutter_appauth_platform_interface: ^3.0.0-nullsafety.0
-  platform_detect: ^1.4.0
-  pointycastle: ^3.0.0-nullsafety.0
+  flutter_appauth_platform_interface: ^3.0.0
+  platform_detect: ^2.0.0
+  pointycastle: ^3.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_appauth: 1.0.0-nullsafety.0
+  flutter_appauth: ^1.0.0+1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
