@@ -6,7 +6,7 @@ class WebClient {
   static const String ERROR_MESSAGE = "[statusCode: %1, description: %2]";
 
   Future<dynamic> post(String url, Map<String, dynamic> body,
-      {Map<String, String> headers}) async {
+      {Map<String, String>? headers}) async {
     try {
       final response =
           await http.post(Uri.parse(url), body: body, headers: headers);
@@ -17,8 +17,11 @@ class WebClient {
         print(jsonResponse["error"].toString());
         throw Exception(ERROR_MESSAGE
             .replaceFirst("%1", response.statusCode.toString())
-            .replaceAll("%2",
-                jsonResponse["error"]?.toString() ?? response.reasonPhrase));
+            .replaceAll(
+                "%2",
+                jsonResponse["error"]?.toString() ??
+                    response.reasonPhrase ??
+                    ""));
       }
       return jsonResponse;
     } catch (e) {
@@ -37,8 +40,11 @@ class WebClient {
       if (response.statusCode != 200) {
         throw Exception(ERROR_MESSAGE
             .replaceFirst("%1", response.statusCode.toString())
-            .replaceAll("%2",
-                jsonResponse["error"]?.toString() ?? response.reasonPhrase));
+            .replaceAll(
+                "%2",
+                jsonResponse["error"]?.toString() ??
+                    response.reasonPhrase ??
+                    ""));
       }
       return jsonResponse;
     } catch (e) {
